Address round-1 review: share data helper, propagate data to client
- Extract undefined_symbol_data_fields() into
  src/utils/undefined-symbol-diagnostic.ts (renamed from
  undefined-symbol-message.ts) and use it at every analyzer→LSP conversion
  (production build_diagnostics + the 5 test harnesses that mirror it),
  removing the hand-copied conditional `data` spread (code-review finding).
- convert_semantic_diagnostic now propagates the structured `data` payload
  onto the published Diagnostic, so symbol_name/reference_kind reach the client
  too (codex blocking finding); previously only the pre-conversion internal
  logic saw it.
- diagnostics-provider test for scalar-vs-variable suppression now attaches
  `data` to its hand-built diagnostic so it exercises the real structured path
  instead of passing vacuously (codex warning).
- string-macro-detection global test now asserts reference_kind === 'global'
  and the $ sigil, restoring the local-vs-global distinction the old wording
  carried implicitly (codex nit).
