+++ tests/property/unify-forward-call-feeds.prop.test.ts
@@ -12,6 +12,7 @@ import { DiagnosticsProvider } from '../../src/providers/diagnostics';
 import { StataLexer } from '../../src/lexer';
 import { StataParser } from '../../src/parser';
 import { SemanticAnalyzer, create_empty_symbol_table } from '../../src/analyzer';
+import { undefined_symbol_data_fields } from '../../src/utils/undefined-symbol-diagnostic';
 import { ContextTracker } from '../../src/context-tracker';
 import { DocumentState, StataDiagnosticCode, ForwardCallSite, StataLSPConfig, ResolvedScope } from '../../src/types';
 import { ScopeResolver } from '../../src/scope-resolver';
@@ -86,9 +87,7 @@ function create_document_state(content: string, uri: string): DocumentState {
             : DiagnosticSeverity.Hint,
         code: diag.code,
         source: 'sight',
-        ...(diag.symbol_name !== undefined || diag.reference_kind !== undefined
-            ? { data: { symbol_name: diag.symbol_name, reference_kind: diag.reference_kind } }
-            : {}),
+        ...undefined_symbol_data_fields(diag),
     }));
 
     return {
